% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R
\name{calc_covparams,PrestoGPModel-method}
\alias{calc_covparams,PrestoGPModel-method}
\title{calc_covparams}
\usage{
\S4method{calc_covparams}{PrestoGPModel}(model, locs, Y)
}
\arguments{
\item{model}{The model to set the covariance parameters of}

\item{locs}{the locations matrix}

\item{Y}{the dependent variable matrix}
}
\value{
a model with initial covariance parameters
}
\description{
Set initial value of covarariance parameters.
}
