% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Vecchia_Spatiotemporal.R
\name{calc_covparams,SpatiotemporalModel-method}
\alias{calc_covparams,SpatiotemporalModel-method}
\title{calc_covparams}
\usage{
\S4method{calc_covparams}{SpatiotemporalModel}(model, locs, Y)
}
\arguments{
\item{model}{The model to set the covariance parameters of}

\item{locs}{the locations matrix}

\item{Y}{the dependent variable matrix}
}
\value{
a model with initial covariance parameters
}
\description{
Set initial value of covarariance parameters.
}
