% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R
\name{compute_error,PrestoGPModel-method}
\alias{compute_error,PrestoGPModel-method}
\title{compute_error}
\usage{
\S4method{compute_error}{PrestoGPModel}(model)
}
\arguments{
\item{model}{the PrestoGP model object}
}
\value{
The total error used in the main optimization loop
}
\description{
Compute the error (log likelihood using the GP log likelihood
and penalty from the beta coefficients)
}
