% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R
\name{compute_residuals,PrestoGPModel-method}
\alias{compute_residuals,PrestoGPModel-method}
\title{compute_residuals}
\usage{
\S4method{compute_residuals}{PrestoGPModel}(model, Y, Y.hat)
}
\arguments{
\item{model}{The model to compute the residual of}

\item{Y}{the training dependent variable matrix}

\item{Y.hat}{the predicted training dependent variable matrix based on beta hat}
}
\value{
a model with computed residuals
}
\description{
Compute residuals based on beta parameters
}
