% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_CreateU_Multivariate.R
\name{create.param.sequence}
\alias{create.param.sequence}
\title{Extract specific Matern parameters from a parameter sequence}
\usage{
create.param.sequence(P, ns = 1)
}
\arguments{
\item{P}{Number of outcomes variables}

\item{ns}{Number of scale parameters}
}
\value{
A matrix with five rows and two columns as described below:
\describe{
\item{Row 1:}{Starting and ending indices for the sigma parameter(s)}
\item{Row 2:}{Starting and ending indices for the scale parameter(s)}
\item{Row 3:}{Starting and ending indices for the smoothness parameter(s)}
\item{Row 4:}{Starting and ending indices for the nugget(s)}
\item{Row 5:}{Starting and ending indices for the correlation parameter(s)}
}
}
\description{
This function is used to obtain specific Matern parameters (e.g.,
range or smoothness) from the covparams slot of a PrestoGPModel object.
}
\examples{
# Space/elevation model
data(soil250, package="geoR")
y2 <- soil250[,7]               # predict pH level
X2 <- as.matrix(soil250[,c(4:6,8:22)])
# columns 1+2 are location coordinates; column 3 is elevation
locs2 <- as.matrix(soil250[,1:3])

soil.vm2 <- new("VecchiaModel", n_neighbors = 10)
# fit separate scale parameters for location and elevation
soil.vm2 <- prestogp_fit(soil.vm2, y2, X2, locs2, scaling = c(1, 1, 2))

pseq <- create.param.sequence(1, 2)
soil2.params <- soil.vm2@covparams
# sigma
soil2.params[pseq[1,1]:pseq[1,2]]
# scale parameters
soil2.params[pseq[2,1]:pseq[2,2]]
# smoothness parameter
soil2.params[pseq[3,1]:pseq[3,2]]
# nugget
soil2.params[pseq[4,1]:pseq[4,2]]

# Multivariate model
data(soil, package="RandomFields")
soil <- soil[!is.na(soil[,5]),] # remove rows with NA's
ym <- list()
ym[[1]] <- soil[,5]             # predict two nitrogen concentration levels
ym[[2]] <- soil[,7]
Xm <- list()
Xm[[1]] <- Xm[[2]] <- as.matrix(soil[,c(4,6,8,9)])
locsm <- list()
locsm[[1]] <- locsm[[2]] <- as.matrix(soil[,1:2])

soil.mvm <-  new("MultivariateVecchiaModel", n_neighbors = 10)
soil.mvm <- prestogp_fit(soil.mvm, ym, Xm, locsm)

pseq <- create.param.sequence(2, 1)
soil.params <- soil.mvm@covparams
# sigmas
soil.params[pseq[1,1]:pseq[1,2]]
# scale parameters
soil.params[pseq[2,1]:pseq[2,2]]
# smoothness parameters
soil.params[pseq[3,1]:pseq[3,2]]
# nuggets
soil.params[pseq[4,1]:pseq[4,2]]
# correlation
soil.corr <- diag(2) / 2
soil.corr[upper.tri(soil.corr)] <- soil.params[pseq[5,1]:pseq[5,2]]
soil.corr <- soil.corr + t(soil.corr)
}
\references{
\itemize{
\item Apanasovich, T.V., Genton, M.G. and Sun, Y. "A valid Matérn class of
cross-covariance functions for multivariate random fields with any number
of components", Journal of the American Statistical Association (2012)
107(497):180-193.
\item Genton, M.G. "Classes of kernels for machine learning: a statistics
perspective", The Journal of Machine Learning Research (2001) 2:299-312.
}
}
\seealso{
\code{\link{PrestoGPModel-class}}
}
