% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R
\name{estimate_betas,PrestoGPModel-method}
\alias{estimate_betas,PrestoGPModel-method}
\title{estimate_betas}
\usage{
\S4method{estimate_betas}{PrestoGPModel}(model, parallel)
}
\arguments{
\item{model}{the model to estimate coeffients for}
}
\value{
A model with updated coefficients
}
\description{
Estimate the beta coefficients for a model (not called by user)
}
