% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Vecchia.R
\name{estimate_theta,VecchiaModel-method}
\alias{estimate_theta,VecchiaModel-method}
\title{estimate_theta}
\usage{
\S4method{estimate_theta}{VecchiaModel}(model, locs, optim.control, method)
}
\arguments{
\item{model}{The model to estimate theta for}

\item{locs}{the locations matrix}
}
\value{
a model with an updated covariance parameters estimate
}
\description{
Estimate covariance parameters during a single round of model optimization
}
