% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Full.R
\docType{class}
\name{FullModel-class}
\alias{FullModel-class}
\alias{FullModel}
\title{Univariate full PrestoGP model class}
\description{
This class is used to create univariate models created with a likelihood
function conditioned on all observations. See \code{\link{PrestoGPModel}}
for a description of the slots.
}
\examples{
pgp.model <- new("FullModel")
}
\seealso{
\code{\link{PrestoGPModel}}
}
