% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R
\name{get_beta,PrestoGPModel-method}
\alias{get_beta,PrestoGPModel-method}
\alias{get_beta}
\title{Extract the regression coefficients (beta) for a PrestoGP model.}
\usage{
\S4method{get_beta}{PrestoGPModel}(model)
}
\arguments{
\item{model}{The PrestoGP model object}
}
\value{
A list containing the regression coefficients. Each element of the
list corresponds to an outcome variable. The final element of the list
contains the intercept(s).
}
\description{
This method extracts a list containing the regression coefficients for
a PrestoGP model.
}
\details{
It is important to note that the intercepts are estimated on the
transformed data. They are not useful for making predictions on the
original (untransformed) data. Use \code{\link{prestogp_predict}} to
make predictions based on a PrestoGP model.
}
\examples{
data(soil)
soil <- soil[!is.na(soil[,5]),] # remove rows with NA's
y <- soil[,4]                   # predict moisture content
X <- as.matrix(soil[,5:9])
locs <- as.matrix(soil[,1:2])

soil.vm <- new("VecchiaModel", n_neighbors = 10)
soil.vm <- prestogp_fit(soil.vm, y, X, locs)
get_beta(soil.vm)
}
\references{
\itemize{
\item Messier, K.P. and Katzfuss, M. "Scalable penalized spatiotemporal
land-use regression for ground-level nitrogen dioxide", The Annals of
Applied Statistics (2021) 15(2):688-710.
}
}
\seealso{
\code{\link{PrestoGPModel-class}}, \code{\link{prestogp_fit}},
\code{\link{prestogp_predict}}
}
