% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_CreateU_Multivariate.R
\name{get_kd_tree_node}
\alias{get_kd_tree_node}
\title{get_kd_tree_node}
\usage{
get_kd_tree_node(kd_tree, location)
}
\arguments{
\item{kd_tree}{A KD tree from the make_kd_tree function}

\item{location}{a location vector with the same number of dimensions as the KD tree}
}
\value{
A vector with the path to the KD tree node (e.g. if the tree pivots are c(0.5, 0.5) and the location is c(0.2, 0.7) this function returns c(1,2) )
}
\description{
Given a KD tree and a location, determine which node of the KD tree has the nearest neighbors to that location.
}
