% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R
\name{get_linear_model,PrestoGPModel-method}
\alias{get_linear_model,PrestoGPModel-method}
\alias{get_linear_model}
\title{Extract the fitted linear model for a PrestoGP model}
\usage{
\S4method{get_linear_model}{PrestoGPModel}(model)
}
\arguments{
\item{model}{The PrestoGP model object}
}
\value{
The fitted linear model (of class cv.glmnet or cv.ncvreg).
}
\description{
This method return the fitted linear model (of class cv.glmnet or cv.ncvreg)
for a PrestoGP model.
}
\details{
It is important to note that the model is fit to the
transformed data. The CV error rate and predicted values of Y will not be
correct for the original (untransformed) data. This method should be
used primarily for examining the coefficient paths and generating plots.
}
\examples{
data(soil)
soil <- soil[!is.na(soil[,5]),] # remove rows with NA's
y <- soil[,4]                   # predict moisture content
X <- as.matrix(soil[,5:9])
locs <- as.matrix(soil[,1:2])

soil.vm <- new("VecchiaModel", n_neighbors = 10)
soil.vm <- prestogp_fit(soil.vm, y, X, locs)
get_linear_model(soil.vm)
}
\references{
\itemize{
\item Messier, K.P. and Katzfuss, M. "Scalable penalized spatiotemporal
land-use regression for ground-level nitrogen dioxide", The Annals of
Applied Statistics (2021) 15(2):688-710.
}
}
\seealso{
\code{\link{PrestoGPModel-class}}, \code{\link{prestogp_fit}},
\code{\link[glmnet]{cv.glmnet}}, \code{\link[ncvreg]{cv.ncvreg}}
}
