% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R
\name{get_theta,PrestoGPModel-method}
\alias{get_theta,PrestoGPModel-method}
\alias{get_theta}
\title{Extract the Matern (theta) parameters from a PrestoGP model.}
\usage{
\S4method{get_theta}{PrestoGPModel}(model)
}
\arguments{
\item{model}{The PrestoGP model object}
}
\value{
A list containing the following elements:
\describe{
\item{sigma:}{A vector containing the estimates of the variance parameter
sigma for each outcome.}
\item{scale:}{A vector containing the estimates of the the scale
parameter(s) for each outcome.}
\item{smoothness:}{A vector containing the estimates of the smoothness
parameter for each outcome.}
\item{nuggets:}{A vector containing the estimates of the nugget variance
for each outcome.}
\item{correlation:}{A matrix containing the estimates of the correlations
between outcomes. Omitted for univariate models.}
}
}
\description{
This method extracts a list containing the Matern variance/covariance
parameters from a PrestoGP model.
}
\examples{
data(soil)
soil <- soil[!is.na(soil[,5]),] # remove rows with NA's
y <- soil[,4]                   # predict moisture content
X <- as.matrix(soil[,5:9])
locs <- as.matrix(soil[,1:2])

soil.vm <- new("VecchiaModel", n_neighbors = 10)
soil.vm <- prestogp_fit(soil.vm, y, X, locs)
get_theta(soil.vm)
}
\references{
\itemize{
\item Messier, K.P. and Katzfuss, M. "Scalable penalized spatiotemporal
land-use regression for ground-level nitrogen dioxide", The Annals of
Applied Statistics (2021) 15(2):688-710.
\item Porcu, E., Bevilacqua, M., Schaback, R., and Oates RJ. "The Matérn
Model: A Journey Through Statistics, Numerical Analysis and Machine
Learning", Statistical Science (2024) 39(3):469-492.
}
}
\seealso{
\code{\link{PrestoGPModel-class}}, \code{\link{prestogp_fit}}
}
