% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R, R/PrestoGP_Vecchia.R,
%   R/PrestoGP_Multivariate_Vecchia.R
\name{get_Y}
\alias{get_Y}
\alias{get_Y,VecchiaModel-method}
\alias{get_Y,MultivariateVecchiaModel-method}
\title{Extract the Y values for a PrestoGP model.}
\usage{
get_Y(model)

\S4method{get_Y}{VecchiaModel}(model)

\S4method{get_Y}{MultivariateVecchiaModel}(model)
}
\arguments{
\item{model}{The PrestoGP model object}
}
\value{
A vector or list containing the values of Y. Any missing Y's will
be replaced with their imputed values.
}
\description{
This method extracts the values of Y for a PrestoGP model. If imputation
was used when fitting the model, the missing Y's will be replaced with
their imputed values.
}
\examples{
data(soil)
soil <- soil[!is.na(soil[,5]),] # remove rows with NA's
y <- soil[,4]                   # predict moisture content
X <- as.matrix(soil[,5:9])
locs <- as.matrix(soil[,1:2])

soil.vm <- new("VecchiaModel", n_neighbors = 10)
soil.vm <- prestogp_fit(soil.vm, y, X, locs)
get_Y(soil.vm)
}
\references{
\itemize{
\item Messier, K.P. and Katzfuss, M. "Scalable penalized spatiotemporal
land-use regression for ground-level nitrogen dioxide", The Annals of
Applied Statistics (2021) 15(2):688-710.
}
}
\seealso{
\code{\link{PrestoGPModel-class}}, \code{\link{prestogp_fit}}
}
