% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Util_Functions.R
\name{glmnet_penalty}
\alias{glmnet_penalty}
\title{glmnet_penalty}
\usage{
glmnet_penalty(beta, lambda, alpha)
}
\arguments{
\item{beta}{the beta coefficients}

\item{lambda}{the penalty factor (how big the total penalty is)}

\item{alpha}{the L1 vs L2 coefficient (0 means all L2, 1 means all L1)}
}
\value{
a penalty score
}
\description{
Compute a combination of L1 and L2 penalties
}
