% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_CreateU_Multivariate.R
\name{knn_indices}
\alias{knn_indices}
\title{knn_indices}
\usage{
knn_indices(ordered_locs, query, n_neighbors, dist_func, dist_func_code)
}
\arguments{
\item{ordered_locs}{A matrix with one row per location, where locations are ordered using max_min ordering.}

\item{query}{Find the nearest neighbors to this location}

\item{n_neighbors}{The number of neighbors to find (K)}

\item{dist_func}{Any distance function with a signature of dist(query_location, locations_matrix)}
}
\value{
A vector containing the indices of the neighbors
}
\description{
Find the index of K nearest neighbors within a set of locations for a given query and distance function
}
