% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_CreateU_Multivariate.R
\name{make_kd_tree}
\alias{make_kd_tree}
\title{make_kd_tree}
\usage{
make_kd_tree(ordered_locs, indices = 1:nrow(ordered_locs), column_idx = 1)
}
\arguments{
\item{ordered_locs}{A matrix with one row per location, where locations are ordered using max_min ordering.}

\item{indices}{A range of indices that should be split into a tree (used for recursion, not by the user)}

\item{column_idx}{The column currently being used to split locations into two branches (used for recursion, not by the user)}
}
\value{
A vector containing the index of locations from sparse to dense
}
\description{
Create a KD tree to help approximate nearest neighbors
}
