% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R
\name{plot_beta,PrestoGPModel-method}
\alias{plot_beta,PrestoGPModel-method}
\alias{plot_beta}
\title{Plots the glide path for the coefficients for a PrestoGP model.}
\usage{
\S4method{plot_beta}{PrestoGPModel}(model, ...)
}
\arguments{
\item{model}{The PrestoGP model object}

\item{...}{Additional parameters to \code{\link[glmnet]{plot.glmnet}} or
\code{\link[ncvreg]{plot.ncvreg}}.}
}
\description{
This method generates a plot showing the coefficients of the model for
different values of the tuning parameter. It is a wrapper for
\code{\link[glmnet]{plot.glmnet}} or \code{\link[ncvreg]{plot.ncvreg}}.
}
\examples{
data(soil)
soil <- soil[!is.na(soil[,5]),] # remove rows with NA's
y <- soil[,4]                   # predict moisture content
X <- as.matrix(soil[,5:9])
locs <- as.matrix(soil[,1:2])

soil.vm <- new("VecchiaModel", n_neighbors = 10)
soil.vm <- prestogp_fit(soil.vm, y, X, locs)
plot_beta(soil.vm)
}
\references{
\itemize{
\item Messier, K.P. and Katzfuss, M. "Scalable penalized spatiotemporal
land-use regression for ground-level nitrogen dioxide", The Annals of
Applied Statistics (2021) 15(2):688-710.
}
}
\seealso{
\code{\link{PrestoGPModel-class}}, \code{\link{prestogp_fit}},
\code{\link[glmnet]{plot.glmnet}}, \code{\link[ncvreg]{plot.ncvreg}}
}
