% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP-package.R
\docType{package}
\name{PrestoGP-package}
\alias{PrestoGP}
\alias{PrestoGP-package}
\title{PrestoGP: Penalized Regression for Spatio-Temporal Outcomes via Gaussian Processes}
\description{
Simultaneous variable seletion and estimation of LUR models with spatiotemporally correlated errors that is scalable for big data.
}
\author{
\strong{Maintainer}: Eric Bair \email{eric.bair@sciome.com}

Authors:
\itemize{
  \item Brian Kidd
  \item Eric Wimberley
  \item Deepak Mav
  \item Kyle Messier
}

}
\keyword{internal}
