% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Multivariate_Spatial.R
\name{prestogp_predict,MultivariateSpatialModel-method}
\alias{prestogp_predict,MultivariateSpatialModel-method}
\title{Make predictions using a previously fit spatiotemporal model.}
\usage{
\S4method{prestogp_predict}{MultivariateSpatialModel}(model, X, locs)
}
\arguments{
\item{model}{A model object returned by prestogp_fit.}

\item{X}{Independent variable matrix for prediction.}

\item{locs}{Locations matrix for which the value is predicted.}
}
\value{
predicted values for the dependent variable
}
\description{
Make predictions using a previously fit spatiotemporal model.
}
\examples{

...
model <- SpatiotemporalModel()
model <- prestogp_fit(model, logNO2, X, locs)
prediction <- prestogp_predict(model, X.test, locs.test)
Vec.mean <- prediction[[1]]
Vec.sds <- prediction[[2]]
}
