% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Vecchia_Spatial.R
\name{prestogp_predict,SpatialModel-method}
\alias{prestogp_predict,SpatialModel-method}
\title{Make predictions using a previously fit spatial model.}
\usage{
\S4method{prestogp_predict}{SpatialModel}(model, X, locs, m = NULL)
}
\arguments{
\item{model}{A model object returned by prestogp_fit.}

\item{X}{Independent variable matrix for prediction.}

\item{locs}{Locations matrix for which the value is predicted.}

\item{m}{The number of neighbors to condition on (if not provided, training value of m will be used).}
}
\value{
predicted values for the dependent variable
}
\description{
Make predictions using a previously fit spatial model.
}
\examples{

...
model <- SpatiotemporalModel()
model <- prestogp_fit(model, logNO2, X, locs)
prediction <- prestogp_predict(model, X.test, locs.test)
Vec.mean <- prediction[[1]]
Vec.sds <- prediction[[2]]
}
