% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R, R/PrestoGP_Vecchia.R,
%   R/PrestoGP_Full.R, R/PrestoGP_Multivariate_Vecchia.R
\name{prestogp_predict}
\alias{prestogp_predict}
\alias{prestogp_predict,VecchiaModel-method}
\alias{prestogp_predict,FullModel-method}
\alias{prestogp_predict,MultivariateVecchiaModel-method}
\title{Prediction for PrestoGP models}
\usage{
prestogp_predict(
  model,
  X = "matrix",
  locs = "matrix",
  m = "numeric",
  ordering.pred = c("obspred", "general"),
  pred.cond = c("independent", "general"),
  return.values = c("mean", "meanvar")
)

\S4method{prestogp_predict}{VecchiaModel}(
  model,
  X = "matrix",
  locs = "matrix",
  m = NULL,
  ordering.pred = c("obspred", "general"),
  pred.cond = c("independent", "general"),
  return.values = c("mean", "meanvar")
)

\S4method{prestogp_predict}{FullModel}(
  model,
  X = "matrix",
  locs = "matrix",
  m = NULL,
  ordering.pred = c("obspred", "general"),
  pred.cond = c("independent", "general"),
  return.values = c("mean", "meanvar")
)

\S4method{prestogp_predict}{MultivariateVecchiaModel}(
  model,
  X = "matrix",
  locs = "matrix",
  m = NULL,
  ordering.pred = c("obspred", "general"),
  pred.cond = c("independent", "general"),
  return.values = c("mean", "meanvar")
)
}
\arguments{
\item{model}{A PrestoGP model object obtained after running
\code{link{prestogp_fit}}.}

\item{X}{The values of the predictor variable(s) for which prediction will
be performed. Should be a matrix for univariate models or a list for
multivariate models.}

\item{locs}{The locations where prediction will be performed. Should be a
matrix for univariate models or a list for multivariate models.}

\item{m}{The number of neighbors to condition on. If not specified, it will
default to the value of m used to fit the model.}

\item{ordering.pred}{Should "obspred" or "general" ordering be used for
prediction? See \code{\link[GPvecchia]{vecchia_specify}} or
\code{\link{vecchia_Mspecify}}. Defaults to "obspred".}

\item{pred.cond}{Should prediction conditioning be "general" or
"indepedent"? See \code{\link[GPvecchia]{vecchia_specify}} or
\code{\link{vecchia_Mspecify}}. Defaults to "independent".}

\item{return.values}{Values that should be returned. Possible values
include "mean" and "meanvar". See
\code{\link[GPvecchia]{vecchia_prediction}} or
\code{\link{vecchia_Mprediction}}. Defaults to "mean".}
}
\value{
A list containing the estimated mean values and (if requested)
standard deviations for the predictions.
}
\description{
After fitting a PrestoGP model, this method can be used to make predictions
on an independent test set (consisting of both new locations and new values
of the predictor variables).
}
\details{
It is important to note that the variance estimates produced by
this function assume that the Matern covariance parameters are fixed and
known. It does not consider any variance resulting from estimating the
covariance parameters or the regression coefficients in PrestoGP models.
These variance estimates will be anticonservative in such cases (and a
warning will be returned when these estimates are calculated).

Prediction is currently not implemented for full models. This function
will return an error if it is applied to a full model.
}
\examples{
data(soil)
soil <- soil[!is.na(soil[,5]),] # remove rows with NA's
y <- soil[,4]                   # predict moisture content
X <- as.matrix(soil[,5:9])
locs <- as.matrix(soil[,1:2])

# Create training and test sets
n <- length(y)
otr <- rep(FALSE, n)
otr[sample(1:n, size=floor(n/2))] <- TRUE
otst <- !otr

# Fit the model on the training set
soil.vm <- new("VecchiaModel", n_neighbors = 10)
soil.vm <- prestogp_fit(soil.vm, y[otr], X[otr,], locs[otr,])

# Perform predictions on the test set
soil.yhat <- prestogp_predict(soil.vm, X[otst,], locs[otst,])
}
\references{
\itemize{
\item Katzfuss, M., and Guinness, J. "A general framework for Vecchia
approximations of Gaussian processes", Statistical Science (2021)
36(1):124-141.
\item Katzfuss, M., Guinness, J., Gong, W. and Zilber, D. "Vecchia
approximations of Gaussian-process predictions", Journal of Agricultural,
Biological and Environmental Statistics (2020) 25:383-414.
\item Messier, K.P. and Katzfuss, M. "Scalable penalized spatiotemporal
land-use regression for ground-level nitrogen dioxide", The Annals of
Applied Statistics (2021) 15(2):688-710.
}
}
\seealso{
\code{\link{PrestoGPModel-class}},\code{\link{prestogp_fit}},
\code{\link[GPvecchia]{vecchia_prediction}},
\code{\link{vecchia_Mprediction}}
}
