% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R
\docType{class}
\name{PrestoGPModel-class}
\alias{PrestoGPModel-class}
\alias{PrestoGPModel}
\title{PrestoGPModel superclass}
\description{
This is the superclass for PrestoGP models. All other types of PrestoGP
model classes (e.g. \code{\link{VecchiaModel-class}} and
\code{\link{FullModel-class}} are inherited from this class. Normally
users should not create objects of this class. Instead, they should use
the appropriate inherited class for the type of model they are fitting.
}
\section{Slots}{

\describe{
\item{\code{covparams}}{A numeric vector containing the parameters for the Matern
model.}

\item{\code{logparams}}{A numeric vector containing the transformed versions of the
Matern parameters (used internally for likelihood calculations).}

\item{\code{beta}}{A column matrix containing the regression coefficients.}

\item{\code{vecchia_approx}}{The output of the Vecchia specify function. See
\code{\link[GPvecchia]{vecchia_specify}} and \code{\link{vecchia_Mspecify}}.}

\item{\code{X_train}}{A matrix containing the original predictors. This will be a
"super matrix" for multivariate models. See
\code{\link[psych]{superMatrix}}.}

\item{\code{Y_train}}{A column matrix containing the original response values.
After fitting the model, missing values will be replaced with the imputed
values if impute.y is TRUE. See \code{link{prestogp_fit}}.}

\item{\code{X_ndx}}{A vector used to find the elements of beta corresponding to
specific outcomes. The \emph{i}th element of X_ndx is the index of the
last element of beta corresponding to predictors for outcome \emph{i}.}

\item{\code{Y_bar}}{A vector containing the means of each outcome.}

\item{\code{Y_obs}}{A logical vector used to track which values of Y are non-missing.}

\item{\code{X_tilde}}{The matrix of transformed predictors.}

\item{\code{y_tilde}}{The column matrix containing the transformed response values.}

\item{\code{res}}{A numeric vector of the residuals.}

\item{\code{locs_train}}{A list containing the location coordinates. Each element
of the list corresponds to a different outcome. (The list will have length
1 for univariate models.)}

\item{\code{linear_model}}{The glmnet model. See \code{\link[glmnet]{glmnet}} and
\code{\link[glmnet]{cv.glmnet}}.}

\item{\code{converged}}{Did the model fitting process converge (boolean)?}

\item{\code{LL_Vecchia_krig}}{The value of the negative log likelihood function
after optimization.}

\item{\code{error}}{Penalized model error. See References for details.}

\item{\code{n_neighbors}}{Number of neighbors to condition on for the Vecchia
approximation. Ignored for full models.}

\item{\code{min_m}}{Minimum permissible number of neighbors.}

\item{\code{alpha}}{Parameter alpha for glmnet. See \code{\link[glmnet]{glmnet}}.}

\item{\code{scaling}}{The indices of the scale parameters. See
\code{link{prestogp_fit}}.}

\item{\code{nscale}}{The number of scale parameters in the model.}

\item{\code{common_scale}}{Do all columns of locs have the same scale parameter?}

\item{\code{param_sequence}}{Records the indices of the various Matern parameters.
See \code{\link{create_param_sequence}}.}
}}

\examples{
pgp.vmodel <- new("VecchiaModel", n_neighbors = 25)
pgp.fmodel <- new("FullModel")
pgp.mmodel <- new("MultivariateVecchiaModel", n_neighbors = 25)

}
\references{
\itemize{
\item Apanasovich, T.V., Genton, M.G. and Sun, Y. "A valid Matérn class of
cross-covariance functions for multivariate random fields with any number
of components", Journal of the American Statistical Association (2012)
107(497):180-193.
\item Messier, K.P. and Katzfuss, M. "Scalable penalized spatiotemporal
land-use regression for ground-level nitrogen dioxide", The Annals of
Applied Statistics (2021) 15(2):688-710.
}
}
\seealso{
\code{\link{VecchiaModel-class}}, \code{\link{FullModel-class}},
\code{\link{MultivariateVecchiaModel-class}}, \code{\link{prestogp_fit}}
}
