% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Vecchia_Spatiotemporal.R
\name{scale_locs,SpatiotemporalModel-method}
\alias{scale_locs,SpatiotemporalModel-method}
\title{scale_locs}
\usage{
\S4method{scale_locs}{SpatiotemporalModel}(model, locs)
}
\arguments{
\item{model}{The model with locations to scale}

\item{locs}{the locations matrix}
}
\value{
a matrix with scaled locations
}
\description{
Scale the locations matrix by the covariance parameters
}
