% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R
\name{show_theta,PrestoGPModel-method}
\alias{show_theta,PrestoGPModel-method}
\title{show_theta}
\usage{
\S4method{show_theta}{PrestoGPModel}(object, Y_names)
}
\arguments{
\item{object}{the PrestoGP model object}

\item{Y_names}{the names of the different outcome variables
(may just be numbers if not provided in training input)}
}
\description{
Print the covariance parameters in a table
}
