% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_CreateU_Multivariate.R
\name{sparseNN}
\alias{sparseNN}
\title{sparseNN}
\usage{
sparseNN(
  ordered_locs,
  n_neighbors,
  dist_func,
  dist_func_code,
  ordered_locs_pred = NULL
)
}
\arguments{
\item{ordered_locs}{A matrix with one row per location, where locations are ordered using max_min ordering.}

\item{n_neighbors}{The number of neighbors to find (K) for each location}

\item{dist_func}{Any distance function with a signature of dist(query_location, locations_matrix)}
}
\value{
A list containing two matrices, each with one row per location: an indices matrix with the indices of nearest neighbors for each location, and a distance matrix with the associated distances
}
\description{
Find the index of K nearest neighbors within a set of locations for a given query and distance function
}
