% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R
\name{sparseToDenseBeta}
\alias{sparseToDenseBeta}
\title{sparseToDenseBeta}
\usage{
sparseToDenseBeta(linear_model)
}
\arguments{
\item{linear_model}{the glmnet model}
}
\value{
A dense matrix
}
\description{
Convert the sparse beta coefficients matrix from glmnet to a dense matrix
}
