% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Vecchia_Spatiotemporal.R
\name{theta_names,SpatiotemporalModel-method}
\alias{theta_names,SpatiotemporalModel-method}
\title{theta_names}
\usage{
\S4method{theta_names}{SpatiotemporalModel}(model)
}
\arguments{
\item{model}{The model to estimate theta for}
}
\value{
a vector with the namems of the covariance parameterss
}
\description{
Return a vector specifying the names of the different covariance parameters (used for show method)
}
