% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Full_Spatial.R
\name{transform_data,FullSpatialModel-method}
\alias{transform_data,FullSpatialModel-method}
\title{transform_data}
\usage{
\S4method{transform_data}{FullSpatialModel}(model, Y, X)
}
\arguments{
\item{model}{A model object.}

\item{Y}{the dependent variable matrix}

\item{X}{the independent variable matrix}
}
\value{
the unmodified model
}
\description{
Model transformation for full models is performed directly on X and Y matrices using the inverted Cholesky decompostion of the covariance matrix.
(no Vecchia approximation)
}
