% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Vecchia.R
\name{transform_data,VecchiaModel-method}
\alias{transform_data,VecchiaModel-method}
\title{transform_data}
\usage{
\S4method{transform_data}{VecchiaModel}(model, Y, X)
}
\arguments{
\item{model}{The model to estimate theta for}

\item{Y}{the dependent variable matrix}

\item{X}{the independent variable matrix}
}
\value{
a model with i.i.d. data
}
\description{
Transform data to be independent, identically distributed
}
