% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Vecchia.R
\docType{class}
\name{VecchiaModel-class}
\alias{VecchiaModel-class}
\alias{VecchiaModel}
\title{Univariate Vecchia PrestoGP model class}
\description{
This class is used to create univariate models with a likelihood function
conditioned on a subset of the observations (i.e., Vecchia models). See
\code{\link{PrestoGPModel}} for a description of the slots.
}
\examples{
pgp.model <- new("VecchiaModel", n_neighbors = 25)
}
\seealso{
\code{\link{PrestoGPModel}}
}
