% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_CreateU_Multivariate.R
\name{max_min_ordering}
\alias{max_min_ordering}
\title{max_min_ordering}
\usage{
max_min_ordering(locs, dist_func)
}
\arguments{
\item{locs}{A matrix with one row per location and any number of columns (x, y, time, etc).}

\item{dist_func}{Any distance function with a signature of dist(query_location, locations_matrix)}
}
\value{
A vector containing the index of locations from sparse to dense
}
\description{
Determine an ordering of locations that uses the max-min method to get a sparse covering of the location space.
}
