% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_Model.R
\name{show,PrestoGPModel-method}
\alias{show,PrestoGPModel-method}
\title{show}
\usage{
\S4method{show}{PrestoGPModel}(object)
}
\arguments{
\item{object}{the PrestoGP model object}
}
\description{
Print a summary of the model and its parameters
}
