% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_Likelihood.R
\name{vecchia_Mlikelihood}
\alias{vecchia_Mlikelihood}
\title{Evaluation of the multivariate Vecchia likelihood}
\usage{
vecchia_Mlikelihood(z, vecchia.approx, covparams)
}
\arguments{
\item{z}{The observed data.}

\item{vecchia.approx}{A Vecchia object returned by
\code{\link{vecchia_Mspecify}}.}

\item{covparams}{Vector of covariance parameters. See
\code{\link{create.param.sequence}} or the examples below for details
about the format of this vector.}
}
\value{
The log likelihood implied by the multivariate Vecchia
approximation.
}
\description{
This function is used to evaluate the multivariate Vecchia likelihood.
}
\examples{
data(soil)
soil <- soil[!is.na(soil[,5]),] # remove rows with NA's
locs <- as.matrix(soil[,1:2])
locsm <- list()
locsm[[1]] <- locsm[[2]] <- locs
soil.va <- vecchia_Mspecify(locsm, m=10)

pseq <- create.param.sequence(2)
# Initialize the vector of covariance parameters
params <- rep(NA, pseq[5,2])
# Sigma parameters:
params[pseq[1,1]:pseq[1,2]] <- c(100, 80)
# Scale parameters:
params[pseq[2,1]:pseq[2,2]] <- c(60, 50)
# Smoothness parameters:
params[pseq[3,1]:pseq[3,2]] <- c(0.5, 0.5)
# Nuggets:
params[pseq[4,1]:pseq[4,2]] <- c(30, 30)
# Correlation:
params[pseq[5,1]:pseq[5,2]] <- -0.9

vecchia_Mlikelihood(rnorm(nrow(locs)), soil.va, params)
}
\references{
\itemize{
\item Katzfuss, M., and Guinness, J. "A general framework for Vecchia
approximations of Gaussian processes", Statistical Science (2021)
36(1):124-141.
}
}
\seealso{
\code{\link[GPvecchia]{vecchia_likelihood}},
\code{\link{vecchia_Mspecify}}, \code{\link{create.param.sequence}}
}
